+++ sdk/netapp/Azure.ResourceManager.NetApp/src/Custom/Models/BackupsMigrationContent.cs
@@ -3,9 +3,8 @@
 
 #nullable disable
 
-using Azure.Core;
-
 using System.ComponentModel;
+using Azure.Core;
 
 namespace Azure.ResourceManager.NetApp.Models
 {