+++ sdk/netapp/Azure.ResourceManager.NetApp/tests/ResourceTests/VolumeGroupTests.cs
@@ -147,7 +147,8 @@ namespace Azure.ResourceManager.NetApp.Tests
             logVolumeProperties.ThroughputMibps = 6;
             logVolumeProperties.ProtocolTypes.InitializeFrom(_nfsProtocolTypes);
             logVolumeProperties.Tags.InitializeFrom(DefaultTags);
-            foreach (var rule in _nfs41ExportPolicyRuleList) { logVolumeProperties.ExportRules.Add(rule); }
+            foreach (var rule in _nfs41ExportPolicyRuleList)
+            { logVolumeProperties.ExportRules.Add(rule); }
             volumeGroupVolumeProperties.Add(logVolumeProperties);
 
             string dataVolumeName = $"{volumeGroupName}-data-1";
@@ -160,7 +161,8 @@ namespace Azure.ResourceManager.NetApp.Tests
             dataVolumeProperties.ThroughputMibps = 6;
             dataVolumeProperties.ProtocolTypes.InitializeFrom(_nfsProtocolTypes);
             dataVolumeProperties.Tags.InitializeFrom(DefaultTags);
-            foreach (var rule in _nfs41ExportPolicyRuleList) { dataVolumeProperties.ExportRules.Add(rule); }
+            foreach (var rule in _nfs41ExportPolicyRuleList)
+            { dataVolumeProperties.ExportRules.Add(rule); }
             volumeGroupVolumeProperties.Add(dataVolumeProperties);
 
             string sharedVolumeName = $"{volumeGroupName}-shared-1";
@@ -173,7 +175,8 @@ namespace Azure.ResourceManager.NetApp.Tests
             sharedVolumeProperties.ThroughputMibps = 6;
             sharedVolumeProperties.ProtocolTypes.InitializeFrom(_nfsProtocolTypes);
             sharedVolumeProperties.Tags.InitializeFrom(DefaultTags);
-            foreach (var rule in _nfs41ExportPolicyRuleList) { sharedVolumeProperties.ExportRules.Add(rule); }
+            foreach (var rule in _nfs41ExportPolicyRuleList)
+            { sharedVolumeProperties.ExportRules.Add(rule); }
             volumeGroupVolumeProperties.Add(sharedVolumeProperties);
 
             IList<NetAppVolumePlacementRule> globalPlacementRules = new List<NetAppVolumePlacementRule> { new NetAppVolumePlacementRule("key1", "value1") };
@@ -181,7 +184,8 @@ namespace Azure.ResourceManager.NetApp.Tests
             NetAppVolumeGroupData volumeGroupDetailsData = new();
             volumeGroupDetailsData.Location = _volumeGroupLocation;
             volumeGroupDetailsData.GroupMetaData = new() { GroupDescription = "group description", ApplicationType = NetAppApplicationType.SAPHANA, ApplicationIdentifier = "SH1" };
-            foreach (var rule in globalPlacementRules) { volumeGroupDetailsData.GroupMetaData.GlobalPlacementRules.Add(rule); }
+            foreach (var rule in globalPlacementRules)
+            { volumeGroupDetailsData.GroupMetaData.GlobalPlacementRules.Add(rule); }
             volumeGroupDetailsData.Volumes.InitializeFrom(volumeGroupVolumeProperties);
 
             NetAppVolumeGroupResource volumeGroupDetails = (await volumeGroupCollection.CreateOrUpdateAsync(WaitUntil.Completed, volumeGroupName, volumeGroupDetailsData)).Value;